feat: refactor on commands
